refactor(test): 문자열 parsing 및 Car 객체 할당 테스트 코드를 분리한다
Game 클래스 내의 기능을 CarAllocator 클래스로 분리함에 따라 테스트 코드도 함께 분리한다.
