+++ __tests__/gameTest.js
@@ -16,26 +16,6 @@ describe('Game 클래스 테스트', () => {
     Random.pickNumberInRange.mockClear();
   });
 
-  describe('parseNames 메서드', () => {
-    test('이름을 구분자로 나누어 배열을 반환한다.', () => {
-      const names = 'Woo,Wa,Han';
-      const result = Game.parseNames(names);
-      expect(result).toEqual(['Woo', 'Wa', 'Han']);
-    });
-  });
-
-  describe('allocateCars 메서드', () => {
-    test('이름 배열을 Car 인스턴스로 변환하여 반환한다.', () => {
-      const nameList = ['Woo', 'Wa', 'Han'];
-      const cars = Game.allocateCars(nameList);
-      expect(cars).toHaveLength(3);
-      expect(cars[0]).toBeInstanceOf(Car);
-      expect(cars[0].name).toBe('Woo');
-      expect(cars[1].name).toBe('Wa');
-      expect(cars[2].name).toBe('Han');
-    });
-  });
-
   describe('play 메서드', () => {
     test('게임을 진행하고 결과를 출력한다.', () => {
       const game = new Game('Woo,Wa', 3);
